Move to enum to meta.
