+++ src/main/java/net/tridentsdk/meta/BarType.java
@@ -1,4 +1,4 @@
-package net.tridentsdk.bar;
+package net.tridentsdk.meta;
 
 /**
  * Enum of all bar types.
